fix: Auth modal 병합
