+++ src/components/auth/index.ts
@@ -1,3 +1,2 @@
-export * from './SignIn';
-export * from './SignUp';
+export * from './Auth';
 export * from './SignOut';
